Refactor Directory
